Rename database setup file
It's no longer just configuration so I think it makes sense to change
the name to something more generic.
